Crate robonomics_rpc

A collection of node-specific RPC methods.

Since substrate core functionality makes no assumptions about the modules used inside the runtime, so do RPC methods defined in sc-rpc crate. It means that client/rpc can’t have any methods that need some strong assumptions about the particular runtime.

The RPCs available in this crate however can make some assumptions about how the runtime is constructed and what FRAME pallets are part of it. Therefore all node-runtime-specific RPCs can be placed here or imported from corresponding FRAME RPC definitions.

Structs

FullDeps

Full client dependencies.

Enums

DenyUnsafe

Signifies whether a potentially unsafe RPC should be denied.

Functions

create_full

Instantiate all Full RPC extensions.

Type Definitions

IoHandler

A IO handler that uses all Full RPC extensions.