Module robonomics_rpc::pubsub

Robonomics Publisher/Subscriber API.

A basic pubsub client demonstrating libp2p and the gossipsub protocol.

Using two terminal windows, start two instances.

target/debug/robonomics --dev --tmp -l rpc=trace
target/debug/robonomics --dev --tmp --ws-port 9991 -l rpc=trace

Then using two terminal windows, start two clients. One of them will send messages, the other one will catch them and print. You can of course open more terminal windows and add more participants.

Pubsub subscribe:

import time
from robonomicsinterface import PubSub, Account

def subscription_handler(obj, update_nr, subscription_id):
    print(obj['params']['result'])
    if update_nr >= 2:
        return 0

account = Account(remote_ws="ws://127.0.0.1:9944")
pubsub = PubSub(account)

print(pubsub.listen("/ip4/127.0.0.1/tcp/44440"))
time.sleep(2)
print(pubsub.connect("/ip4/127.0.0.1/tcp/44441"))
time.sleep(2)
print(pubsub.subscribe("topic_name", result_handler=subscription_handler))

Pubsub publish:

import time
from robonomicsinterface import PubSub

account = Account(remote_ws="ws://127.0.0.1:9991")
pubsub = PubSub(account)

print(pubsub.listen("/ip4/127.0.0.1/tcp/44441"))
time.sleep(2)
print(pubsub.connect("/ip4/127.0.0.1/tcp/44440"))
time.sleep(2)

for i in range(10):
    time.sleep(2)
    print("publish:", pubsub.publish("topic_name", "message_" + str(time.time())))

Structs

PubSubRpc

Traits

PubSubRpcServer

Server trait implementation for the PubSubRpc RPC API.