Trait robonomics_rpc::pubsub::PubSubRpcServer

pub trait PubSubRpcServer: Sized + Send + Sync + 'static {
    fn peer_id(&self) -> RpcResult<String>;
    fn listen<'life0, 'async_trait>(
        &'life0 self,
        address: Multiaddr
    ) -> Pin<Box<dyn Future<Output = RpcResult<bool>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn listeners<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Multiaddr>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn connect<'life0, 'async_trait>(
        &'life0 self,
        address: Multiaddr
    ) -> Pin<Box<dyn Future<Output = RpcResult<bool>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn publish<'life0, 'async_trait>(
        &'life0 self,
        topic_name: String,
        message: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<bool>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn subscribe(
        &self,
        subscription_sink: SubscriptionSink,
        topic_name: String
    ) -> SubscriptionResult;

    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the PubSubRpc RPC API.

Required Methods

fn peer_id(&self) -> RpcResult<String>

Returns local peer ID.

fn listen<'life0, 'async_trait>(
&'life0 self,
address: Multiaddr
) -> Pin<Box<dyn Future<Output = RpcResult<bool>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Listen address for incoming connections.

fn listeners<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Multiaddr>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a list of node addresses.

fn connect<'life0, 'async_trait>(
&'life0 self,
address: Multiaddr
) -> Pin<Box<dyn Future<Output = RpcResult<bool>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Connect to peer and add it into swarm.

fn publish<'life0, 'async_trait>(
&'life0 self,
topic_name: String,
message: String
) -> Pin<Box<dyn Future<Output = RpcResult<bool>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Publish message into the topic by name.

fn subscribe(
&self,
subscription_sink: SubscriptionSink,
topic_name: String
) -> SubscriptionResult

Subscribe for a topic with given name.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.